chore: Refactor DBF to SQL Converter to support multiple files
Refactor the `convert.js` file to accept a folder path instead of a single file path. This allows the conversion of multiple DBF files in a specified directory to SQL files. The converted SQL files are saved in an output folder with the same name as the DBF file. Each DBF file is converted into a SQL script that creates a table and inserts the data from the DBF file.

This change improves the usability and flexibility of the DBF to SQL Converter.